Fix some possible implicit casting warnings where "strtol"/"strtoul" is used

+++ source/effect_codegen_hlsl.cpp
@@ -466,7 +466,7 @@ private:
 				digit_index--;
 			digit_index++;
 
-			const uint32_t semantic_digit = std::strtoul(semantic.c_str() + digit_index, nullptr, 10);
+			const uint32_t semantic_digit = static_cast<uint32_t>(std::strtoul(semantic.c_str() + digit_index, nullptr, 10));
 			const std::string semantic_base = semantic.substr(0, digit_index);
 
 			if (semantic_base == "TEXCOORD")
